requirements.txt
streamlit
selenium
pytesseract
pillow
xgboost
pandas
